+++ views/src/components/NavBar.tsx
@@ -34,7 +34,7 @@ const NavBar: FC = () => {
     return (
         <Fragment>
             <ReactIf condition={isAuthenticated}>
-                <Navbar expand='lg' style={{ zoom: 0.85 }} fixed='top'>
+                <Navbar variant='dark' expand='lg' style={{ zoom: 0.85 }} fixed='top'>
                     <Container>
                         <Link href='/dataplatform'><Navbar.Brand style={{ fontSize: '1.3rem' }}>Lenstack</Navbar.Brand></Link>
                         <Navbar.Toggle aria-controls='navbarScroll' />
@@ -59,7 +59,7 @@ const NavBar: FC = () => {
                 </Navbar>
             </ReactIf>
             <ReactIf condition={!isAuthenticated}>
-                <Navbar className='navbar-unauthorized' expand='lg'>
+                <Navbar variant='dark' className='navbar-unauthorized' expand='lg'>
                     <Container>
                         <Link href='/'>
                             <Navbar.Brand style={{ fontSize: '1.3rem' }}>Lenstack</Navbar.Brand>