+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Lenstack Changelog
 
+## [Currrent v2.7]
+* New UI of view dataset page.
+* Added dataset ratings.
+* Added card in dataset page.
+
 ## [Currrent v2.6]
 * Fixed issues.
 * New sorting algorithm.